chore: 메타데이터 적용 #45
