+++ src/app/layout.tsx
@@ -11,8 +11,27 @@ import { Toaster } from "@/components/common/sonner"
 import { cn } from "@/utils/cn"
 
 export const metadata: Metadata = {
-  title: "Create Next App",
-  description: "Generated by create next app"
+  title: {
+    default: "피키토키",
+    template: "피키토키 | %s"
+  },
+  description: "친구들과 우리만의 채널을 공유하고 소통해보세요!",
+  openGraph: {
+    siteName: "피키토키",
+    title: "피키토키",
+    type: "website",
+    description: "친구들과 우리만의 채널을 공유하고 소통해보세요!",
+    images: [
+      {
+        url: "/logo/picki_talki_logo_on_white.png",
+        alt: "Picki-Talki OG Image",
+        width: 940,
+        height: 194
+      }
+    ],
+    url: "https://dnd-12th-4-frontend.vercel.app"
+  },
+  metadataBase: new URL("https://dnd-12th-4-frontend.vercel.app")
 }
 
 export default async function RootLayout({
