+++ src/components/channels/ChannelBoxForm.tsx
@@ -4,6 +4,7 @@ import Tag from "@/components/common/Tag"
 import { cn } from "@/utils/cn"
 import { share, ShareAPIRequest } from "@/utils/share"
 import { Check, Share2 } from "lucide-react"
+import Image from "next/image"
 import { useRouter } from "next/navigation"
 import { useFormContext } from "react-hook-form"
 interface Params {
@@ -89,7 +90,10 @@ export default function ChannelBoxForm({
       <article className="flex justify-between text-[12px] text-black/60">
         {/* <p>{memberCount}명과 소통 중이에요.</p> */}
         {/* <p>채널장 : {owner}</p> */}
-        <p className="text-caption-02 text-emphasis-medium">응답을 기다리는 중..</p>
+        <div className="flex items-center">
+          <Image src={"/clock.png"} alt="clock-icon" width={24} height={24} />
+          <p className="text-caption-02 text-emphasis-medium">응답을 기다리는 중..</p>
+        </div>
         <p className="text-caption-01 text-emphasis-medium">
           <span className="text-secondary-300">{memberCount}명</span>과 소통 중이에요.
         </p>
